requirements.txt
aiosqlite==0.20.0
httpx==0.27.2
pydantic==2.9.2
PyYAML==6.0.2
python-dotenv==1.0.1
tenacity==9.0.0
